requirements.txt
streamlit
pdfplumber
pymupdf
google-generativeai
python-dotenv
Pillow
jinja2
fpdf2
